+++ src/main/java/domain/deck/Rank.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.deck;
 
 public enum Rank {
     ACE("A", 1),
