+++ src/main/java/domain/deck/Suit.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.deck;
 
 public enum Suit {
     SPADE("스페이드"), DIAMOND("다이아몬드"), HEART("하트"), CLOVER("클로버");
