+++ src/main/java/domain/player/Dealer.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 public class Dealer extends Player {
     private static final String DEALER_NAME = "딜러";
