+++ src/main/java/domain/player/Hand.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.player;
 
+import domain.deck.Card;
+import domain.deck.Rank;
 import java.util.LinkedList;
 import java.util.List;
