+++ src/main/java/domain/player/Name.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 public class Name {
     private static final int MIN_LENGTH = 1;
